+++ Demo/DemoTests/MoyaProviderIntegrationTests.swift
@@ -90,7 +90,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
                 describe("a provider with credential plugin") {
                     it("credential closure returns nil") {
                         var called = false
-                        let plugin = CredentialsPlugin<HTTPBin> { (target) -> (NSURLCredential?) in
+                        let plugin = CredentialsPlugin { (target) -> (NSURLCredential?) in
                             called = true
                             return nil
                         }
@@ -108,7 +108,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
                     it("credential closure returns valid username and password") {
                         var called = false
                         var returnedData: NSData?
-                        let plugin = CredentialsPlugin<HTTPBin> { (target) -> (NSURLCredential?) in
+                        let plugin = CredentialsPlugin { (target) -> (NSURLCredential?) in
                             called = true
                             return NSURLCredential(user: "user", password: "passwd", persistence: .None)
                         }
@@ -127,7 +127,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
                 describe("a provider with network activity plugin") {
                     it("notifies at the beginning of network requests") {
                         var called = false
-                        let plugin = NetworkActivityPlugin<GitHub> { (change) -> () in
+                        let plugin = NetworkActivityPlugin { (change) -> () in
                             if change == .Began {
                                 called = true
                             }
@@ -142,7 +142,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
 
                     it("notifies at the end of network requests") {
                         var called = false
-                        let plugin = NetworkActivityPlugin<GitHub> { (change) -> () in
+                        let plugin = NetworkActivityPlugin { (change) -> () in
                             if change == .Ended {
                                 called = true
                             }