+++ Demo/DemoTests/ReactiveCocoaMoyaProviderTests.swift
@@ -106,7 +106,7 @@ class ReactiveCocoaMoyaProviderSpec: QuickSpec {
                     requestClosure: RequestClosure = MoyaProvider.DefaultRequestMapping,
                     stubClosure: StubClosure = MoyaProvider.NeverStub,
                     manager: Manager = Alamofire.Manager.sharedInstance,
-                    plugins: [Plugin<Target>] = []) {
+                    plugins: [Plugin] = []) {
 
                         super.init(endpointClosure: endpointClosure, requestClosure: requestClosure, stubClosure: stubClosure, manager: manager, plugins: plugins)
                 }
@@ -188,7 +188,7 @@ class ReactiveCocoaMoyaProviderSpec: QuickSpec {
                         requestClosure: RequestClosure = MoyaProvider.DefaultRequestMapping,
                         stubClosure: StubClosure = MoyaProvider.NeverStub,
                         manager: Manager = Alamofire.Manager.sharedInstance,
-                        plugins: [Plugin<Target>] = []) {
+                        plugins: [Plugin] = []) {
 
                             super.init(endpointClosure: endpointClosure, requestClosure: requestClosure, stubClosure: stubClosure, manager: manager, plugins: plugins)
                     }