+++ Source/Moya.swift
@@ -92,14 +92,14 @@ public class MoyaProvider<Target: MoyaTarget> {
     
     /// A list of plugins
     /// e.g. for logging, network activity indicator or credentials
-    public let plugins: [Plugin<Target>]
+    public let plugins: [Plugin]
 
     /// Initializes a provider.
     public init(endpointClosure: EndpointClosure = MoyaProvider.DefaultEndpointMapping,
         requestClosure: RequestClosure = MoyaProvider.DefaultRequestMapping,
         stubClosure: StubClosure = MoyaProvider.NeverStub,
         manager: Manager = Alamofire.Manager.sharedInstance,
-        plugins: [Plugin<Target>] = []) {
+        plugins: [Plugin] = []) {
 
         self.endpointClosure = endpointClosure
         self.requestClosure = requestClosure
@@ -202,14 +202,14 @@ internal extension MoyaProvider {
         let plugins = self.plugins
         
         // Give plugins the chance to alter the outgoing request
-        plugins.forEach { $0.willSendRequest(request, provider: self, target: target) }
+        plugins.forEach { $0.willSendRequest(request, target: target) }
         
         // Perform the actual request
         let alamoRequest = request.response { (_, response: NSHTTPURLResponse?, data: NSData?, error: NSError?) -> () in
             let statusCode = response?.statusCode
 
             // Inform all plugins about the response
-            plugins.forEach { $0.didReceiveResponse(data, statusCode: statusCode, response: response, error: error, provider: self, target: target) }
+            plugins.forEach { $0.didReceiveResponse(data, statusCode: statusCode, response: response, error: error, target: target) }
             completion(data: data, statusCode: statusCode, response: response, error: error)
         }
         
@@ -218,21 +218,21 @@ internal extension MoyaProvider {
     }
 
     /// Creates a function which, when called, executes the appropriate stubbing behavior for the given parameters.
-    internal final func createStubFunction(token: CancellableToken, forTarget target: Target, withCompletion completion: Moya.Completion, endpoint: Endpoint<Target>, plugins: [Plugin<Target>]) -> (() -> ()) {
+    internal final func createStubFunction(token: CancellableToken, forTarget target: Target, withCompletion completion: Moya.Completion, endpoint: Endpoint<Target>, plugins: [Plugin]) -> (() -> ()) {
         return {
             if (token.canceled) {
                 let error = NSError(domain: NSURLErrorDomain, code: NSURLErrorCancelled, userInfo: nil)
-                plugins.forEach { $0.didReceiveResponse(nil, statusCode: nil, response: nil, error: error, provider: self, target: target) }
+                plugins.forEach { $0.didReceiveResponse(nil, statusCode: nil, response: nil, error: error,target: target) }
                 completion(data: nil, statusCode: nil, response: nil, error: error)
                 return
             }
 
             switch endpoint.sampleResponseClosure() {
             case .NetworkResponse(let statusCode, let data):
-                plugins.forEach { $0.didReceiveResponse(data, statusCode: statusCode, response: nil, error: nil, provider: self, target: target) }
+                plugins.forEach { $0.didReceiveResponse(data, statusCode: statusCode, response: nil, error: nil, target: target) }
                 completion(data: data, statusCode: statusCode, response: nil, error: nil)
             case .NetworkError(let error):
-                plugins.forEach { $0.didReceiveResponse(nil, statusCode: nil, response: nil, error: error, provider: self, target: target) }
+                plugins.forEach { $0.didReceiveResponse(nil, statusCode: nil, response: nil, error: error, target: target) }
                 completion(data: nil, statusCode: nil, response: nil, error: error)
             }
         }
@@ -241,7 +241,7 @@ internal extension MoyaProvider {
     /// Notify all plugins that a stub is about to be performed. You must call this if overriding `stubRequest`.
     internal final func notifyPluginsOfImpendingStub(request: NSURLRequest, target: Target) {
         let request = manager.request(request)
-        plugins.forEach { $0.willSendRequest(request, provider: self, target: target) }
+        plugins.forEach { $0.willSendRequest(request, target: target) }
     }
 }