+++ Changelog.md
@@ -1,5 +1,6 @@
 # Next
 
+- **Breaking Change** `Plugin<Target>` is now a protocol and as such no longer sends a typed `MoyaProvider`. - @swizzlr
 - **Breaking Change** `MoyaResponse` is now `Response`, and also `final`. It will be changed to a `struct` in a future release. - @swizzlr
 - **Breaking Change** `ReactiveCocoaMoyaProvider` can now be supplied with an optional `stubScheduler` – @swizzlr (sponsored by [Network Locum](https://networklocum.com))
 - **Breaking Change** Introduce `MoyaError` type for use with reactive extensions - [@tomburns](http://github.com/tomburns)